requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
motor==3.6.0
neo4j==5.27.0
pydantic-settings==2.7.1
python-dotenv==1.0.1
